backend_python/requirements.txt
supabase
fastapi 
uvicorn
playwright
python-dotenv
sqlalchemy 
psycopg2-binary
